LayoutTests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: line 10: TypeError: Argument 1 ('node') to Node.appendChild must be an instance of Node
